Move compatibility code to the shared/compatibility.js.
