+++ examples/components/simpleviewer.html
@@ -37,9 +37,6 @@ limitations under the License.
 
   <link rel="stylesheet" href="../../build/dist/web/pdf_viewer.css">
 
-  <!-- for legacy browsers -->
-  <script src="../../build/dist/web/compatibility.js"></script>
-
   <script src="../../build/dist/build/pdf.js"></script>
   <script src="../../build/dist/web/pdf_viewer.js"></script>
 </head>
