+++ examples/learning/helloworld64.html
@@ -10,9 +10,6 @@
 
 <canvas id="the-canvas" style="border:1px  solid black"></canvas>
 
-<!-- for legacy browsers we need to use compatibility.js -->
-<script src="../../web/compatibility.js"></script>
-
 <script src="../../build/generic/build/pdf.js"></script>
 
 <script id="script">
